pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meistroverse"
dynamic = ["version"]
description = "Omni-Behavioral Evolutionary Logic Interface for Synchronic Knowledge"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Mark", email = "mark@meistroverse.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.23",
    "pymysql>=1.1.0",
    "redis>=5.0.1",
    "faiss-cpu>=1.7.4",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "openai>=1.3.7",
    "anthropic>=0.8.1",
    "sentence-transformers>=2.2.2",
    "celery>=5.3.4",
    "requests>=2.31.0",
    "httpx>=0.25.2",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "python-json-logger>=2.0.7",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/meistroverse/meistroverse"
Documentation = "https://github.com/meistroverse/meistroverse"
Repository = "https://github.com/meistroverse/meistroverse"

[tool.hatch.version]
path = "meistroverse/__init__.py"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88